needed to merge with develop to get this to work correctly
